+++ app/src/types/match_data.hpp
@@ -36,7 +36,7 @@ struct MoveData {
     int seldepth           = 0;
     int depth              = 0;
     int score              = 0;
-    int nps                = 0;
+    uint64_t nps           = 0;
     int hashfull           = 0;
     uint64_t tbhits        = 0;
     bool legal             = true;
